Refactored and hashed out more of the skeleton
