+++ provider-directory/src/main/java/gov/va/api/health/providerdirectory/service/controller/endpoint/EndpointTransformer.java
@@ -1,0 +1,4 @@
+package gov.va.api.health.providerdirectory.service.controller.endpoint;
+
+public class EndpointTransformer {
+}
