+++ provider-directory/src/main/java/gov/va/api/health/providerdirectory/service/controller/greeting/GreetingController.java
@@ -1,6 +1,7 @@
-package gov.va.api.health;
+package gov.va.api.health.providerdirectory.service.controller.greeting;
 
 import java.util.concurrent.atomic.AtomicLong;
+
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
