+++ provider-directory/src/main/java/gov/va/api/health/providerdirectory/service/controller/location/LocationController.java
@@ -1,0 +1,4 @@
+package gov.va.api.health.providerdirectory.service.controller.location;
+
+public class LocationController {
+}
