+++ provider-directory/src/main/java/gov/va/api/health/providerdirectory/service/controller/organization/OrganizationTransformer.java
@@ -1,0 +1,4 @@
+package gov.va.api.health.providerdirectory.service.controller.organization;
+
+public class OrganizationTransformer {
+}
