+++ provider-directory/src/main/java/gov/va/api/health/providerdirectory/service/controller/practitioner/PractitionerTransformer.java
@@ -1,0 +1,4 @@
+package gov.va.api.health.providerdirectory.service.controller.practitioner;
+
+public class PractitionerTransformer {
+}
